+++ extension/chrome/elements/composer/composer-abstract-component.ts
