+++ extension/chrome/elements/composer/composer-render.ts
@@ -3,19 +3,18 @@
 'use strict';
 
 import { ComposerComponent } from './composer-abstract-component.js';
-import { Recipients } from './composer-types.js';
-import { Ui, JQS } from '../browser.js';
-import { RecipientType } from '../api/api.js';
-import { Xss } from '../platform/xss.js';
-import { Catch } from '../platform/catch.js';
-import { BrowserMsg } from '../extension.js';
-import { Lang } from '../lang.js';
-import { KeyImportUi } from '../ui/key_import_ui.js';
-import { Pgp } from '../core/pgp.js';
-import { Str } from '../core/common.js';
-import { Store } from '../platform/store.js';
-import { SendableMsg } from '../api/email_provider/email_provider_api.js';
-import { Att } from '../core/att.js';
+import { Ui, JQS } from '../../../js/common/browser.js';
+import { RecipientType } from '../../../js/common/api/api.js';
+import { Xss } from '../../../js/common/platform/xss.js';
+import { Catch } from '../../../js/common/platform/catch.js';
+import { BrowserMsg } from '../../../js/common/extension.js';
+import { Lang } from '../../../js/common/lang.js';
+import { KeyImportUi } from '../../../js/common/ui/key_import_ui.js';
+import { Pgp } from '../../../js/common/core/pgp.js';
+import { Str } from '../../../js/common/core/common.js';
+import { Store } from '../../../js/common/platform/store.js';
+import { SendableMsg, Recipients } from '../../../js/common/api/email_provider/email_provider_api.js';
+import { Att } from '../../../js/common/core/att.js';
 
 export class ComposerRender extends ComposerComponent {
