+++ extension/chrome/elements/composer/composer-send-btn-popover.ts
@@ -4,11 +4,11 @@
 
 import { ComposerComponent } from './composer-abstract-component.js';
 import { PopoverOpt, PopoverChoices } from './composer-types.js';
-import { Xss } from '../platform/xss.js';
-import { Lang } from '../lang.js';
-import { Ui } from '../browser.js';
-import { Catch } from '../platform/catch.js';
-import { Store } from '../platform/store.js';
+import { Xss } from '../../../js/common/platform/xss.js';
+import { Lang } from '../../../js/common/lang.js';
+import { Ui } from '../../../js/common/browser.js';
+import { Catch } from '../../../js/common/platform/catch.js';
+import { Store } from '../../../js/common/platform/store.js';
 
 export class ComposerSendBtnPopover extends ComposerComponent {
