+++ extension/chrome/elements/composer/formatters/composer-mail-formatter.ts
@@ -3,12 +3,12 @@
 'use strict';
 
 import { NewMsgData } from "../composer-types.js";
-import { KeyInfo } from "../../core/pgp.js";
+import { KeyInfo } from "../../../../js/common/core/pgp.js";
 import { Composer } from "../composer.js";
 import { PlainMsgMailFormatter } from './plain-mail-msg-formatter.js';
 import { SignedMsgMailFormatter } from './signed-msg-mail-formatter.js';
 import { EncryptedMsgMailFormatter } from './encrypted-mail-msg-formatter.js';
-import { SendableMsg } from '../../api/email_provider/email_provider_api.js';
+import { SendableMsg } from '../../../../js/common/api/email_provider/email_provider_api.js';
 
 export class GeneralMailFormatter {
